backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
python-socketio==5.11.0
psycopg2-binary==2.9.9
GeoAlchemy2==0.14.2
Shapely==2.0.2
numpy<2
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
pyotp==2.9.0
requests==2.31.0
gunicorn==21.2.0
twilio==9.0.4
